chore: move context and client to shared
